Support union types in property promotion
