+++ tests/Integration/Bridge/TolerantParser/Reflection/ReflectionPropertyTest.php
@@ -16,7 +16,6 @@ class ReflectionPropertyTest extends IntegrationTestCase
     /**
      * @dataProvider provideReflectionPropertyTypes
      * @dataProvider provideReflectionProperty
-     * @dataProvider provideConsturctorPropertyPromotion
      */
     public function testReflectProperty(string $source, string $class, \Closure $assertion): void
     {
@@ -369,58 +368,4 @@ EOT
             ];
     }
 
-
-    public function provideConsturctorPropertyPromotion(): Generator
-    {
-        yield 'Typed properties' => [
-                <<<'EOT'
-<?php
-
-namespace Test;
-
-class Barfoo
-{
-    public function __construct(
-        private string $foobar
-        private int $barfoo
-    ) {}
-}
-EOT
-                ,
-                'Test\Barfoo',
-                function (ReflectionPropertyCollection $properties) {
-                    $this->assertTrue($properties->get('foobar')->isPromoted());
-                    $this->assertEquals(
-                        Type::string(),
-                        $properties->get('foobar')->type()
-                    );
-                    $this->assertEquals(
-                        Type::int(),
-                        $properties->get('barfoo')->type()
-                    );
-                },
-            ];
-
-        yield 'Nullable' => [
-                '<?php class Barfoo { public function __construct(private ?string $foobar){}}',
-                'Barfoo',
-                function (ReflectionPropertyCollection $properties) {
-                    $this->assertEquals(
-                        Type::string()->asNullable(),
-                        $properties->get('foobar')->type()
-                    );
-                },
-            ];
-
-        yield 'No types' => [
-                '<?php class Barfoo { public function __construct(private $foobar){}}',
-                'Barfoo',
-                function (ReflectionPropertyCollection $properties) {
-                    $this->assertEquals(
-                        Type::undefined(),
-                        $properties->get('foobar')->type()
-                    );
-                },
-            ];
-    }
 }
